fix: unopened cart,add refresh state after redirect,fix params in cart page
